+++ src/hooks/websocketListener/useInitializeWebsocketConnection.tsx
@@ -23,17 +23,15 @@ const DISCONNECT = 'disconnect';
 export function useInitializeWebsocketConnection() {
   const messageTimeout = useRef<NodeJS.Timeout | null>(null);
   const batchTimeout = useRef<NodeJS.Timeout | null>(null);
-
   const { address } = useGetAccount();
-
   const dispatch = useDispatch();
-
   const { network } = useGetNetworkConfig();
 
   const handleMessageReceived = (message: string) => {
     if (messageTimeout.current) {
       clearTimeout(messageTimeout.current);
     }
+
     messageTimeout.current = setTimeout(() => {
       dispatch(setWebsocketEvent(message));
     }, MESSAGE_DELAY);
@@ -43,14 +41,29 @@ export function useInitializeWebsocketConnection() {
     if (batchTimeout.current) {
       clearTimeout(batchTimeout.current);
     }
+
     batchTimeout.current = setTimeout(() => {
       dispatch(setWebsocketBatchEvent(data));
     }, MESSAGE_DELAY);
   };
 
+  const unsubscribeWS = () => {
+    websocketConnection.current?.close();
+    websocketConnection.current = null;
+    websocketConnection.status = WebsocketConnectionStatusEnum.NOT_INITIALIZED;
+
+    if (messageTimeout.current) {
+      clearTimeout(messageTimeout.current);
+    }
+  };
+
   const initializeWebsocketConnection = useCallback(
     retryMultipleTimes(
       async () => {
+        if (!address) {
+          return;
+        }
+
         // If there are many components that use this hook, the initialize method is triggered many times.
         // To avoid multiple connections to the same endpoint, we have to guard the initialization before the logic started
         websocketConnection.status = WebsocketConnectionStatusEnum.PENDING;
@@ -82,11 +95,18 @@ export function useInitializeWebsocketConnection() {
         });
 
         websocketConnection.current.on(DISCONNECT, () => {
-          console.warn('Websocket disconnected. Trying to reconnect...');
-          setTimeout(() => {
-            console.log('Websocket reconnecting...');
-            websocketConnection.current?.connect();
-          }, RETRY_INTERVAL);
+          if (address) {
+            // Make sure we are still logged in before retrying to connect to the websocket
+            console.warn('Websocket disconnected. Trying to reconnect...');
+
+            setTimeout(() => {
+              if (address) {
+                // Make sure we are still logged in when the timeout is finished
+                console.log('Websocket reconnecting...');
+                websocketConnection.current?.connect();
+              }
+            }, RETRY_INTERVAL);
+          }
         });
       },
       {
@@ -105,17 +125,18 @@ export function useInitializeWebsocketConnection() {
       !websocketConnection.current?.active
     ) {
       initializeWebsocketConnection();
+      return;
+    }
+
+    if (!address) {
+      // Close the websocket connection when we are not logged in
+      unsubscribeWS();
     }
   }, [address, initializeWebsocketConnection]);
 
   useEffect(() => {
     return () => {
-      websocketConnection.current?.close();
-      websocketConnection.status =
-        WebsocketConnectionStatusEnum.NOT_INITIALIZED;
-      if (messageTimeout.current) {
-        clearTimeout(messageTimeout.current);
-      }
+      unsubscribeWS();
     };
   }, []);
 }